Update test for new child folder behavior
