+++ packages/components/src/internal/containerFixtures.ts
@@ -1,6 +1,9 @@
 import { Project } from '@labkey/api';
 
 import { Container, ContainerDateFormats } from './components/base/models/Container';
+import { AppContext } from './AppContext';
+import { getTestAPIWrapper } from './APIWrapper';
+import { getSecurityTestAPIWrapper } from './components/security/APIWrapper';
 
 const TEST_DATE_FORMATS: ContainerDateFormats = {
     dateFormat: 'yyyy-MM-dd',
@@ -72,3 +75,17 @@ export const TEST_FOLDER_OTHER_CONTAINER_ADMIN = new Container({
     ...TEST_FOLDER_OTHER_CONTAINER_CONFIG,
     effectivePermissions: ['org.labkey.api.security.permissions.AdminPermission'],
 });
+export const TEST_PROJECT_APP_CONTEXT_ADMIN: Partial<AppContext> = {
+    api: getTestAPIWrapper(jest.fn, {
+        security: getSecurityTestAPIWrapper(jest.fn, {
+            fetchContainers: jest.fn().mockResolvedValue([TEST_PROJECT_CONTAINER_ADMIN, TEST_FOLDER_CONTAINER_ADMIN]),
+        }),
+    }),
+};
+export const TEST_PROJECT_APP_CONTEXT_NON_ADMIN: Partial<AppContext> = {
+    api: getTestAPIWrapper(jest.fn, {
+        security: getSecurityTestAPIWrapper(jest.fn, {
+            fetchContainers: jest.fn().mockResolvedValue([TEST_PROJECT_CONTAINER, TEST_FOLDER_CONTAINER]),
+        }),
+    }),
+};
